refactor: migrate network-retrofit to use shared network module
