--- core/network-retrofit/src/main/java/com/nisrulz/example/spacexapi/network/retrofit/RemoteDataSource.kt
@@ -1,7 +1,0 @@
-package com.nisrulz.example.spacexapi.network.retrofit
-
-import com.nisrulz.example.spacexapi.network.retrofit.dto.LaunchInfoResponse
-
-interface RemoteDataSource {
-    suspend fun getAllLaunches(): List<LaunchInfoResponse>
-}
