--- core/network-retrofit/src/main/java/com/nisrulz/example/spacexapi/network/retrofit/dto/Crew.kt
@@ -1,9 +1,0 @@
-package com.nisrulz.example.spacexapi.network.retrofit.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Crew(
-    val crew: String = "",
-    val role: String = ""
-)
