--- core/network-retrofit/src/main/java/com/nisrulz/example/spacexapi/network/retrofit/dto/Failure.kt
@@ -1,10 +1,0 @@
-package com.nisrulz.example.spacexapi.network.retrofit.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Failure(
-    val altitude: Int = 0,
-    val reason: String = "",
-    val time: Int = 0
-)
