--- core/network-retrofit/src/main/java/com/nisrulz/example/spacexapi/network/retrofit/dto/Fairings.kt
@@ -1,11 +1,0 @@
-package com.nisrulz.example.spacexapi.network.retrofit.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Fairings(
-    val recovered: Boolean = false,
-    val recovery_attempt: Boolean = false,
-    val reused: Boolean = false,
-    val ships: List<String>
-)
