--- core/network-retrofit/src/main/java/com/nisrulz/example/spacexapi/network/retrofit/dto/Patch.kt
@@ -1,9 +1,0 @@
-package com.nisrulz.example.spacexapi.network.retrofit.dto
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class Patch(
-    val large: String = "",
-    val small: String = ""
-)
